redux imports refactored, setUserData bug
