+++ src/redux/actions/homePage.action.js
@@ -1,5 +1,10 @@
-import {SET_CURRENT_PAGE, SET_TENDERS, SET_TOTAL_COUNT, TOGGLE_IS_LOADED} from "../actionTypes/homePage.actionType";
-import {getTendersAPI} from "../../services/homePage.services";
+import {
+    SET_CURRENT_PAGE,
+    SET_TENDERS,
+    SET_TOTAL_COUNT,
+    TOGGLE_IS_LOADED
+} from "@redux/actionTypes";
+import {getTendersAPI} from "@services";
 
 export const setTenders = (tenders) => ({type: SET_TENDERS, tenders})
 export const setTotalCount = (totalCount) => ({type: SET_TOTAL_COUNT, totalCount})
@@ -13,8 +18,6 @@ export const getHomeTenders = (currentPage = 1, pageSize = 10) => async (dispatc
     const tenders = await getTendersAPI(currentPage, pageSize)
 
     dispatch(setTenders(tenders.data))
-    // dispatch(setTotalCount(data.totalCount))
     dispatch(setTotalCount(tenders.data.length))
     dispatch(toggleIsLoaded(true))
-
 }
