+++ src/redux/actions/index.js
@@ -1,0 +1,5 @@
+export * from './accountPage.action'
+export * from './accountTender.action'
+export * from './homePage.action'
+export * from './homeTender.action'
+export * from './auth.action'
