+++ src/selectors/auth.selector.js
