+++ src/selectors/index.js
@@ -1,0 +1,3 @@
+export * from './accountPage.selector'
+export * from './auth.selector'
+export * from './homePage.selector'
