+++ src/services/index.js
@@ -1,0 +1,6 @@
+import AuthenticationService from './auth.service'
+export default AuthenticationService
+
+export * from './accountPage.service'
+export * from './auth.service'
+export * from './homePage.service'
